Have a single place for pulling events
