+++ packages/ai-bot/lib/matrix.ts
@@ -1,4 +1,4 @@
-import { IContent } from 'matrix-js-sdk';
+import { IContent, Method } from 'matrix-js-sdk';
 import { logger } from '@cardstack/runtime-common';
 import { OpenAIError } from 'openai/error';
 import * as Sentry from '@sentry/node';
@@ -11,6 +11,8 @@ import {
   APP_BOXEL_REASONING_CONTENT_KEY,
   APP_BOXEL_MESSAGE_MSGTYPE,
 } from '@cardstack/runtime-common/matrix-constants';
+import type { MatrixEvent as DiscreteMatrixEvent } from 'https://cardstack.com/base/matrix-event';
+import { encodeUri } from 'matrix-js-sdk/lib/utils';
 
 let log = logger('ai-bot');
 
@@ -146,6 +148,24 @@ function cleanupCache() {
       fileCache.delete(url);
     }
   }
+}
+
+export async function getRoomEvents(roomId: string, client: MatrixClient) {
+  const messagesPath = encodeUri('/rooms/$roomId/messages', {
+    $roomId: roomId,
+  });
+  let response = await (client as any).http.authedRequest(
+    Method.Get,
+    messagesPath,
+    {
+      dir: 'f',
+      limit: '1000',
+      filter: JSON.stringify({
+        'org.matrix.msc3874.not_rel_types': ['m.replace'],
+      }),
+    },
+  );
+  return (response?.chunk || []) as DiscreteMatrixEvent[];
 }
 
 export async function downloadFile(
